[SIG-3480] Cypress split tests
