+++ e2e-tests/package.json
@@ -2,7 +2,7 @@
   "name": "@signalen/e2e-tests",
   "private": true,
   "dependencies": {
-    "cypress": "^6.8.0",
+    "cypress": "^7.1.0",
     "jsonwebtoken": "^8.5.1"
   },
   "scripts": {
